move test util functions to test_util.go
